+++ hub20/apps/ethereum/models/transfers.py
@@ -3,7 +3,6 @@ import logging
 from django.db import models
 
 from hub20.apps.core.models import (
-    PaymentNetwork_T,
     TokenAmount,
     Transfer,
     TransferConfirmation,
@@ -11,7 +10,6 @@ from hub20.apps.core.models import (
     TransferReceipt,
 )
 
-from .accounts import EthereumAccount_T
 from .blockchain import Transaction, TransactionDataRecord
 from .fields import EthereumAddressField
 from .networks import BlockchainPaymentNetwork
@@ -20,14 +18,14 @@ logger = logging.getLogger(__name__)
 
 
 class BlockchainTransfer(Transfer):
-    NETWORK: PaymentNetwork_T = BlockchainPaymentNetwork
+    NETWORK = BlockchainPaymentNetwork
     address = EthereumAddressField(db_index=True)
 
     def _execute(self):
         try:
             assert self.provider is not None, "No active provider to execute transfer"
 
-            account: EthereumAccount_T = self.provider.select_for_transfer(self.as_token_amount)
+            account = self.provider.select_for_transfer(self.as_token_amount)
             assert account is not None, "No account with enough balance to cover"
             tx_data: TransactionDataRecord = self.provider.transfer(
                 amount=self.as_token_amount, address=self.address